feat: :fire: Cleaned up repetitive types in the interfaces
