+++ src/models/genreData.d.ts
@@ -1,3 +1,7 @@
 export interface IGenreData {
+  id: number;
   genreName: string;
+  slug: string;
+  comic_count: number;
+  group: string;
 }
